Init
Copy-Paste of incomplete/in progress tutorial from original private repository. Currently awaiting feedback/assistance from my institution regarding networking, as well as review/validation of security procedures, but this process is taking longer than expected. I expect this should be sufficient for the needs of most people, and/or point them in the correct direction for the time being. Updates will be forthcoming, assuming I hear back at all. In the meant time, if you're reading this and have relevant knowledge, please don't hesitate to contribute!